moved test file to a test folder, so tests will run on GitHub
